In case you were concerned about the docs being too excellent
I have no idea what this parameter does, is supposed to do, or anything. None doesn't work. This does. 

Same for the question of why I have an empty menu. Because having items in the menu is confusing, and having no menu makes it invisible. I do not understand, and am willing to ignore the warnings caused when you try to look at an empty menu.